+++ packages/support/src/mixins/index.ts
@@ -1,0 +1,12 @@
+import Builder from "./Builder";
+export {
+    Builder
+};
+
+export * from './decorators';
+export * from './apply';
+export * from './hasMixin';
+export * from './isApplicationOf';
+export * from './mix';
+export * from './unwrap';
+export * from './wrap';