+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+* `@aedart/support/mixins` submodule that offers an adaptation of Justin Fagnani's [`mixwith.js`](https://github.com/justinfagnani/mixwith.js).
+* `NOTICE` files for the root package, and all provided packages.
+
+### Changed
+
+* Generic type is now optional (_defaults to `object`_) for `Constructor`, `AbstractConstructor` and `ConstructorOrAbstractConstructor` types, in `@aedart/contracts/support`.
+
 ## [0.7.0] - 2024-02-02
 
 ### Added